refactor: structure  new model of portfolios and division of responsabilities
