+++ public/components/header.ejs
@@ -4,7 +4,7 @@
     <nav class="menu">
         <ul>
             <a href="/"><li>Home Page</li></a>
-            <a href="profile"><li>Profile</li></a>
+            <a href="/signin"><li>SignIn</li></a>
             <a href="/about"><li>About</li></a>
         </ul>
     </nav>
