+++ public/signin.ejs
@@ -4,12 +4,12 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Profile</title>
+    <title>Sign In</title>
 </head>
 <body>
     
     <%- include ('components/header.ejs') %>
-    <%- include ('components/login.ejs') %>
+    <%- include ('components/signin.ejs') %>
 
 
 </body>
